+++ NEWS.md
@@ -1,4 +1,11 @@
-# rotor 0.3.1
+# rotor 0.3.2
+
+* Hotfix for some tests related to the `Cache` R6 class that fail on systems
+  with low-precision file system timestamps (such as ext3 and old Windows
+  file systems)
+
+
+# rotor 0.3.2
 
 * fixes time zone related issue in `Cache$prune()`.
 * fixes bug in `rotate_rds(on_change_only = TRUE)` that occurs if a version